MAINT: Partial refactor of diffTpfa
