+++ tests/numerics/fv/test_ad_tpfa.py
@@ -86,204 +86,6 @@ class AdTpfaFlow(pp.fluid_mass_balance.SinglePhaseFlow):
         # return pp.wrap_as_dense_ad_array(scaled_vals, name="permeability") + e_xy @ p
         return pp.wrap_as_dense_ad_array(np.ones(nc * tensor_dim), name="permeability")
 
-    def _block_diagonal_grid_property_matrix(
-        self,
-        domains: list[pp.Grid],
-        grid_property_getter: Callable[[pp.Grid], Any],
-        name: Optional[str] = None,
-    ) -> sps.spmatrix:
-        """Construct mapping matrix for the connectivity between two grids entities.
-
-        The mapping matrix is a block diagonal matrix where each block contains 1 where
-        the two entities are connected, and 0 otherwise.
-
-        Parameters:
-            domains: List of grids.
-            grid_property_getter: Function that returns the property of the grid that
-                should be used for the mapping.
-            name: Name of the operator.
-
-        Returns:
-            Mapping matrix.
-        """
-        blocks = []
-        for g in domains:
-            if g.dim == 0:
-                # 0d subdomains have no faces, so the projection might not exist.
-                # TODO: Implement this with proper handling of special cases (first and
-                # second dimension being faces and cells, respectively).
-                raise NotImplementedError
-            else:
-                mat_loc = grid_property_getter(g)
-            blocks.append(mat_loc)
-
-        block_matrix = pp.matrix_operations.optimized_compressed_storage(
-            sps.block_diag(blocks)
-        )
-        return block_matrix
-
-    def cells_to_half_faces(
-        self, subdomains: list[pp.Grid], dim: int, with_sign=False
-    ) -> sps.spmatrix:
-        """Mapping from cells to half-faces.
-
-        Parameters:
-            subdomains: List of grids.
-            dim: Dimension of the half-faces.
-
-        Returns:
-            Operator.
-        """
-
-        def get_matrix(g: pp.Grid) -> sps.csr_matrix:
-            _, ci, sgn = sps.find(g.cell_faces)
-            # Repeat dim times in f order
-            row_inds = np.repeat(np.arange(ci.size), dim)
-            col_inds = pp.fvutils.expand_indices_nd(ci, dim)
-            if with_sign:
-                vals = np.repeat(sgn, dim)
-            else:
-                vals = np.ones(col_inds.size)
-            mat = sps.csr_matrix(
-                (vals, (row_inds, col_inds)),
-                shape=(ci.size, g.num_cells * dim),
-            )
-            return mat
-
-        return self._block_diagonal_grid_property_matrix(
-            subdomains,
-            get_matrix,
-        )
-
-    def faces_to_half_faces(
-        self, subdomains: list[pp.Grid], dim: int, with_sign=False
-    ) -> sps.spmatrix:
-        """Mapping from faces to half-faces.
-
-        Parameters:
-            subdomains: List of grids.
-            dim: Dimension of the half-faces.
-
-        Returns:
-            Operator.
-        """
-
-        def get_matrix(g: pp.Grid) -> sps.csr_matrix:
-            fi, _, sgn = sps.find(g.cell_faces)
-
-            # Repeat dim times in f order
-            row_inds = np.repeat(np.arange(fi.size), dim)
-            col_inds = pp.fvutils.expand_indices_nd(fi, dim)
-            if with_sign:
-                vals = np.repeat(sgn, dim)
-            else:
-                vals = np.ones(col_inds.size)
-            mat = sps.csr_matrix(
-                (vals, (row_inds, col_inds)),
-                shape=(fi.size, g.num_faces * dim),
-            )
-            return mat
-
-        return self._block_diagonal_grid_property_matrix(
-            subdomains,
-            get_matrix,
-        )
-
-    def cell_face_vectors(self, subdomains: list[pp.Grid]) -> sps.spmatrix:
-        """Distance between face centers and cell centers.
-
-        Parameters:
-            subdomains: List of grids.
-
-        Returns:
-            Operator.
-        """
-
-        vec_dim = 3
-
-        def get_c_f_vec_matrix(g: pp.Grid) -> sps.csr_matrix:
-            """Construct matrix of vectors connecting cell centers and face centers."""
-
-            fi, ci, sgn = sps.find(g.cell_faces)
-            # Construct vectors from cell centers to face centers.
-            fc_cc = g.face_centers[:, fi] - g.cell_centers[:, ci]
-            # Repeat dim times in f order
-            num_hf = fi.size
-            # Each row contains vec_dim entries and corresponds to one half-face.
-            row_inds = np.repeat(np.arange(num_hf), vec_dim)
-            # There are num_hf * vec_dim columns, each vec_dim-long block corresponding
-            # to one half-face.
-            col_inds = pp.fvutils.expand_indices_nd(np.arange(num_hf), vec_dim)
-            vals = fc_cc.ravel("F")
-            mat = sps.csr_matrix(
-                (vals, (row_inds, col_inds)),
-                shape=(num_hf, num_hf * vec_dim),
-            )
-            return mat
-
-        dist_vec = self._block_diagonal_grid_property_matrix(
-            subdomains,
-            get_c_f_vec_matrix,
-        )
-
-        return dist_vec
-
-    def normal_vectors(self, subdomains: list[pp.Grid]) -> sps.spmatrix:
-        """Normal vectors on half-faces, repeated for each dimension.
-
-        Parameters:
-            subdomains: List of grids.
-
-        Returns:
-            Operator.
-        """
-        vector_dim = 3
-
-        def get_matrix(g: pp.Grid) -> sps.csr_matrix:
-            """Construct normal vector matrix. Each vector is repeated vector_dim times.
-
-            Half-face i corresponds to rows
-                vector_dim * i:vector_dim(i+1)
-            and contains n_0^i, n_1^i, n_2^i. The column indices makes sure we hit the
-            right permeability entries. The permeability being a tensor_dim * num_cells
-            vector, we expand the cell indices to tensor_dim indices.
-
-            Parameters:
-                g: Grid.
-
-            Returns:
-                spmatrix ``(num_half_faces * vector_dim, num_cells * tensor_dim)``:
-                    Normal vector matrix.
-
-            """
-            tensor_dim = vector_dim**2
-            fi, ci, sgn = sps.find(g.cell_faces)
-            num_hf = fi.size
-            n = g.face_normals
-            row_inds = np.repeat(np.arange(num_hf * vector_dim), vector_dim)
-            col_inds = pp.fvutils.expand_indices_nd(ci, tensor_dim)
-            repeat_fi = np.repeat(fi, vector_dim)
-            vals = n[:, repeat_fi].ravel("F")
-            mat = sps.csr_matrix(
-                (vals, (row_inds, col_inds)),
-                shape=(num_hf * vector_dim, g.num_cells * vector_dim**2),
-            )
-            return mat
-
-        return self._block_diagonal_grid_property_matrix(
-            subdomains,
-            get_matrix,
-        )
-
-    def cell_face_distances(self, subdomains: list[pp.Grid]) -> sps.spmatrix:
-        """Scalar distance between face centers and cell centers for each half face."""
-        vals = []
-        for g in subdomains:
-            fi, ci, sgn = sps.find(g.cell_faces)
-            fc_cc = g.face_centers[:, fi] - g.cell_centers[:, ci]
-            vals.append(np.power(fc_cc, 2).sum(axis=0))
-        return np.hstack(vals)
-
     def tpfa_ad(self, subdomains: list[pp.Grid]) -> pp.ad.Operator:
         """TPFA discretization.
 
@@ -293,31 +95,10 @@ class AdTpfaFlow(pp.fluid_mass_balance.SinglePhaseFlow):
         Returns:
             Operator.
         """
-        # Half-face transmissibilities are computed as
-        # t_hf = d_vec @ n @ k_hf / dist
-        # k_hf: Permeability on half-faces shape=(9 x num_half_faces,)
-        # n: Normal vector on half-faces shape=(3 x num_half_faces,9 x num_half_faces)
-        # d_vec: Vectors cell centers and face centers shape=(num_half_faces, 3 x num_half_faces)
-        # dist: Distance between cell centers and face centers shape=(num_half_faces,)
-        d_vec = self.cell_face_vectors(subdomains)
-        dist = self.cell_face_distances(subdomains)
-        k_c = self._permeability(subdomains)
-        n = self.normal_vectors(subdomains)
+        discr = pp.numerics.fv.tpfa.DifferentiableTpfa(self.mdg, subdomains=subdomains)
 
-        d_n_by_dist = sps.diags(1 / dist) * d_vec @ n
-        one = pp.ad.Scalar(1)
-        # t_hf = d_vec @ n @ k_hf / dist
-        t_hf_inv = one / (pp.ad.SparseArray(d_n_by_dist) @ k_c)
-        # Sum over half-faces to get transmissibility on faces.
-        # Include sign to cancel the effect of the d_vec @ n having opposite signs on
-        # the two half-faces.
-        hf_to_f_mat = self.faces_to_half_faces(subdomains, dim=1, with_sign=True).T
-        T_f = one / (pp.ad.SparseArray(hf_to_f_mat) @ t_hf_inv)
-
-        # Construct difference operator to get p_l - p_r on faces. First map p to half-
-        # faces, then to faces with the signed matrix.
-        c_to_hf_mat = self.cells_to_half_faces(subdomains, dim=1, with_sign=False)
-        diff_p = pp.ad.SparseArray(hf_to_f_mat @ c_to_hf_mat) @ self.pressure(
+        T_f = discr.face_transmissibilities(subdomains)
+        diff_p = discr.face_pairing_from_cell_vector(subdomains) @ self.pressure(
             subdomains
         )
 
@@ -343,11 +124,7 @@ class AdTpfaFlow(pp.fluid_mass_balance.SinglePhaseFlow):
         )
         proj = pp.ad.BoundaryProjection(self.mdg, subdomains=subdomains, dim=1)
 
-        # Construct sign vector on boundary from the signed hf_to_f_mat. The following
-        # pairing of signs from half faces to faces works because there's only one half
-        # face at each boundary face. Internally, the sign contributions cancel.
-        one_vec = np.ones(hf_to_f_mat.shape[1])
-        bnd_sgn = pp.ad.DenseArray(hf_to_f_mat @ one_vec)
+        bnd_sgn = discr.boundary_sign(subdomains)
         # Neu values in T_bf = bnd_sgn.
         neu_bnd = proj.boundary_to_subdomain @ neu_filter * bnd_sgn
         # Dir values in T_bf = -bnd_sgn * T_f
